喜欢动画

+++ app/src/main/java/com/d6/android/app/dialogs/OpenDateDialog.kt
@@ -80,7 +80,7 @@ class OpenDateDialog : DialogFragment(),RequestManager {
         dismissAllowingStateLoss()
         isBaseActivity {
             //194ecdb4-4809-4b2d-bf32-42a3342964df
-            Request.signUpdate(userId.toInt(),myAppointment?.sId.toString(),"").request(it,success = { msg, data ->
+            Request.signUpdate(userId,myAppointment?.sId.toString(),"").request(it,success = { msg, data ->
                 var openSuccessDialog = OpenDateSuccessDialog()
                 openSuccessDialog.show(it.supportFragmentManager, "d")
             }) { code, msg ->
